fix(stock_notification_service) Fixing the Kafka handle_sneaker event handler
- Added an explicit event type check for calling the delete_sneaker coroutine

+++ backend/stock_notification_service/stock_notification_service/stock_notification/kafka/kafka_handlers/sneaker_handler.py
@@ -25,7 +25,7 @@ async def handle_sneaker(key: str | None, value: dict):
                 sneaker_update = SneakerUpdate.model_validate(data, strict=False)
                 await update_sneaker(session, sneaker_id, sneaker_update)
 
-            else:
+            elif event_type == "sneaker_deleted":
                 sneaker_id = value.get("sneaker_id")
                 await delete_sneaker(session, sneaker_id)
     except Exception as e:
